wip: first idea for pipelines
